Refactor island environment components
