+++ src/game/scenery/island/environment/Sea.ts
@@ -1,12 +1,83 @@
 import {each, every} from 'lodash';
 import { WORLD_SIZE } from '../../../../utils/lba';
-import { applyLightningUniforms } from './lightning';
 import * as THREE from 'three';
 import { compile } from '../../../../utils/shaders';
 import VERT_SEA from './shaders/sea.vert.glsl';
 import FRAG_SEA from './shaders/sea.frag.glsl';
 import { loadSubTexture, makeNoiseTexture } from '../../../../texture';
 import { getLightVector } from '../geometries';
+import Lightning from './Lightning';
+
+const worldScale = 1 / (WORLD_SIZE * 0.04);
+
+export default class Sea {
+    threeObject: THREE.Object3D;
+    private uniforms: any;
+
+    constructor(props, data, envInfo, usedTiles, layout) {
+        const positions = [];
+        each(layout.seaSections, (section) => {
+            const xd = Math.floor(section.x / 2);
+            const zd = Math.floor(section.z / 2);
+            const offsetX = 1 - Math.abs(section.x % 2);
+            const offsetZ = Math.abs(section.z % 2);
+            const tilesKey = [xd, zd].join(',');
+            loadSeaGeometry(
+                section,
+                positions,
+                usedTiles[tilesKey],
+                offsetX,
+                offsetZ,
+                envInfo.index
+            );
+        });
+
+        const light = getLightVector(data.ambience);
+        const noiseTexture = makeNoiseTexture('LBA_SEA');
+
+        this.uniforms = {
+            uTexture: {
+                value: loadSubTexture(
+                    data.ress.getEntry(envInfo.index),
+                    data.palette,
+                    0,
+                    0,
+                    128,
+                    128
+                )
+            },
+            fogColor: {value: new THREE.Vector3().fromArray(envInfo.skyColor)},
+            fogDensity: {value: envInfo.fogDensity},
+            worldScale: {value: worldScale},
+            time: {value: 0.0},
+            scale: {value: props.scale},
+            light: {value: light},
+            noise: {value: noiseTexture},
+            amplitude: {value: props.amplitude || 0.1}
+        };
+
+        const material = new THREE.RawShaderMaterial({
+            vertexShader: compile('vert', VERT_SEA),
+            fragmentShader: compile('frag', FRAG_SEA),
+            uniforms: this.uniforms,
+            // wireframe: true
+        });
+
+        const bufferGeometry = new THREE.BufferGeometry();
+        bufferGeometry.setAttribute(
+            'position',
+            new THREE.BufferAttribute(new Float32Array(positions), 3)
+        );
+        this.threeObject = new THREE.Mesh(bufferGeometry, material);
+        this.threeObject.matrixAutoUpdate = false;
+        this.threeObject.name = name;
+        this.threeObject.onBeforeRender = Lightning.applyUniforms;
+    }
+
+    update(_game, _scene, time) {
+        this.uniforms.time.value = time.elapsed;
+    }
+}
 
 const push = Array.prototype.push;
 
@@ -64,68 +135,6 @@ const triangles = {
         [[0.5, 1], [1, 0.5], [1, 1]]
     ]
 };
-
-const worldScale = 1 / (WORLD_SIZE * 0.04);
-
-export function loadSea(props, {layout, usedTiles, envInfo, ress, palette, ambience}) {
-    const positions = [];
-    each(layout.seaSections, (section) => {
-        const xd = Math.floor(section.x / 2);
-        const zd = Math.floor(section.z / 2);
-        const offsetX = 1 - Math.abs(section.x % 2);
-        const offsetZ = Math.abs(section.z % 2);
-        const tilesKey = [xd, zd].join(',');
-        loadSeaGeometry(
-            section,
-            positions,
-            usedTiles[tilesKey],
-            offsetX,
-            offsetZ,
-            envInfo.index
-        );
-    });
-
-    const light = getLightVector(ambience);
-    const noiseTexture = makeNoiseTexture('LBA_SEA');
-
-    const uniforms = {
-        uTexture: {
-            value: loadSubTexture(ress.getEntry(envInfo.index), palette, 0, 0, 128, 128)
-        },
-        fogColor: {value: new THREE.Vector3().fromArray(envInfo.skyColor)},
-        fogDensity: {value: envInfo.fogDensity},
-        worldScale: {value: worldScale},
-        time: {value: 0.0},
-        scale: {value: props.scale},
-        light: {value: light},
-        noise: {value: noiseTexture},
-        amplitude: {value: props.amplitude || 0.1}
-    };
-
-    const material = new THREE.RawShaderMaterial({
-        vertexShader: compile('vert', VERT_SEA),
-        fragmentShader: compile('frag', FRAG_SEA),
-        uniforms,
-        // wireframe: true
-    });
-
-    const bufferGeometry = new THREE.BufferGeometry();
-    bufferGeometry.setAttribute(
-        'position',
-        new THREE.BufferAttribute(new Float32Array(positions), 3)
-    );
-    const seaMesh = new THREE.Mesh(bufferGeometry, material);
-    seaMesh.matrixAutoUpdate = false;
-    seaMesh.name = name;
-    seaMesh.onBeforeRender = applyLightningUniforms;
-
-    return {
-        threeObject: seaMesh,
-        update: (_game, _scene, time) => {
-            uniforms.time.value = time.elapsed;
-        },
-    };
-}
 
 function loadSeaGeometry(section, positions, usedTile, offsetX, offsetZ, skyIndex) {
     const n = Math.pow(2, 2 - section.lod) * 8;
